+++ client/src/get-ticket/get-ticket.css
@@ -1,5 +1,3 @@
-/* NAVBAR */
-
 .navbar {
     width: 100%;
     position: absolute;
@@ -43,7 +41,6 @@
     color: white;
 }
 
-/* Italian Post Office */
 .post-office {
     display: flex;
     align-items: center;
@@ -62,7 +59,6 @@
     text-align: left;
 }
 
-/* Ticket Details */
 .ticket-details {
     text-align: center;
     font-weight: bold;
@@ -70,10 +66,8 @@
     font-size: 40px;
     margin-bottom: 10px;
     margin-top: 50px;
-    
 }
 
-/* Ticket Number Box */
 .ticket-box {
     border: 2px solid orange;
     padding-top: 20px;
@@ -85,23 +79,19 @@
     width: 250px;
 }
 
-/* Ticket Text */
 .ticket-text {
     color: #2b2b2b;
     font-size: 24px;
     text-align: center;
 }
 
-/* Ticket Number */
 .ticket-number {
     text-align: center;
     color: orange;
     font-weight: bold;
     font-size: 100px;
-    
 }
 
-/* Information Box */
 .information-box {
     display: flex;
     justify-content: space-between;
@@ -109,7 +99,6 @@
     width: 500px;
 }
 
-/* waiting Time */
 .waiting-time {
     width: 45%;
     text-align: center;
@@ -117,14 +106,12 @@
     font-size: 24px;
 }
 
-/* Service Type */
 .service-type {
     width: 45%;
     color: #103783;
     font-size: 24px;
 }
 
-/* waiting Time number */
 .waiting-time-number {
     text-align: center;
     color: #2b2b2b;
@@ -133,7 +120,6 @@
     padding-top: 5px;
 }
 
-/* Service Type number */
 .service-type-number {
     text-align: center;
     color: #2b2b2b;
@@ -142,16 +128,12 @@
     font-size: 20px;
 }
 
-
-/* Waiting Time and Service Type*/
 .wait-service {
     display: flex;
     justify-content: space-around;
     text-align: center;
 }
 
-
-/* Get Your Receipt */
 .get-receipt {
     margin-top: 20px;
 }
@@ -163,7 +145,6 @@
     color: #103783;
 }
 
-/* Button Container */
 .button-container {
     display: flex;
     flex-wrap: wrap;
@@ -192,7 +173,6 @@
     font-weight: bold;
 }
 
-/* Help button*/
 .help-box {
     margin-top: 50px;        
 }
@@ -209,4 +189,3 @@
     color: #103783;
     border-color: #103783;
 }
-
