+++ client/src/intefaces/types.ts
